Fix GitHub Actions URL when hosting on a different repo
When building for GitHub Pages hosting on a different repository, the
GitHub repo address needs to point to where the tests were run.

This commit adds a `github_tests_repo` input that defaults to repository
on which the action is run, instead of mapping to the `github_repo`
input. This should work automatically when a different host repo is
used, and allow customization for any other needs.

Signed-off-by: Mike Szczys <szczys@hotmail.com>